refactor(tests): quarantine V2.01 parity tests under tests/legacy_v201/
Move tests that pin V2.01 (legacy TOYO_Origin_2.01.py) compatibility
into a dedicated tests/legacy_v201/ subdirectory to clarify the boundary
between primary-spec and legacy-compatibility tests. Pure file
relocation — test count and coverage unchanged (250 passed).

- Move tests/test_parity.py to tests/legacy_v201/test_parity.py
- Extract the seven V2.01 PTN-dialect-handling tests from
  tests/test_reader.py into tests/legacy_v201/test_reader_v201_quirks.py
- Add tests/legacy_v201/__init__.py and tests/legacy_v201/README.md
  documenting the quarantine boundary and pointer to issue #105
- CHANGELOG: record the move under [Unreleased] / Changed (#95)

Closes #95
