+++ CHANGELOG.md
@@ -22,6 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   extended `STATE_JA_TO_EN` (and the new reverse `STATE_EN_TO_JA`) with
   the `充電休止` / `放電休止` substates emitted by native `連続データ.csv`
   exports. Pure refactor — no runtime behaviour change. ([#93])
+- Moved V2.01 parity tests into `tests/legacy_v201/` to clarify the
+  boundary between primary-spec and legacy-compatibility tests. ([#95])
 
 ## [0.1.8] - 2026-04-27
 
@@ -292,6 +294,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#86]: https://github.com/tomooki/toyo-battery/pull/86
 [#88]: https://github.com/tomooki/toyo-battery/pull/88
 [#93]: https://github.com/tomooki/toyo-battery/issues/93
+[#95]: https://github.com/tomooki/toyo-battery/issues/95
 [#99]: https://github.com/tomooki/toyo-battery/issues/99
 [#107]: https://github.com/tomooki/toyo-battery/pull/107
 [#108]: https://github.com/tomooki/toyo-battery/pull/108
