+++ tests/legacy_v201/README.md
@@ -1,0 +1,3 @@
+# `tests/legacy_v201/`
+
+These tests pin V2.01 (legacy `TOYO_Origin_2.01.py`) compatibility for parity reasons. They are NOT the primary specification for `echemplot` behavior — see `tests/test_*.py` at the parent level for that. When V2.01 parity is consciously broken (e.g. as part of the 0.2.0 cleanup tracked by issue #105), update or remove these tests rather than letting them drift.
